# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Todolist App API",
    "version": "1.0.0",
    "description": "Actual REST API contract for the current backend implementation."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Invalid request." },
          "details": {
            "type": "array",
            "items": { "type": "object", "additionalProperties": true }
          }
        },
        "required": ["code", "message"]
      },
...
      "TodoStatus": {
        "type": "string",
        "enum": ["PLANNED", "IN_PROGRESS", "DONE", "ON_HOLD"]
      },
      "Todo": {
        "type": "object",
        "properties": {
          "todoId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid", "nullable": true },
          "teamId": { "type": "string", "format": "uuid", "nullable": true },
          "categoryId": { "type": "string", "format": "uuid", "nullable": true },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "status": { "$ref": "#/components/schemas/TodoStatus" },
          "startDate": { "type": "string", "format": "date-time", "nullable": true },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["todoId", "title", "status", "createdAt", "updatedAt"]
      },
      "TodoCreateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "maxLength": 500 },
          "description": { "type": "string" },
          "status": { "$ref": "#/components/schemas/TodoStatus" },
          "startDate": { "type": "string", "format": "date" },
          "dueDate": { "type": "string", "format": "date" },
          "categoryId": { "type": "string", "format": "uuid", "nullable": true },
          "teamId": { "type": "string", "format": "uuid", "nullable": true }
        },
        "required": ["title"]
      },
      "TodoUpdateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "maxLength": 500 },
          "description": { "type": "string" },
          "status": { "$ref": "#/components/schemas/TodoStatus" },
          "startDate": { "type": "string", "format": "date", "nullable": true },
          "dueDate": { "type": "string", "format": "date", "nullable": true },
          "categoryId": { "type": "string", "format": "uuid", "nullable": true }
        }
      },
      "TodoStatusUpdateRequest": {
        "type": "object",
        "properties": {
          "status": { "$ref": "#/components/schemas/TodoStatus" }
        },
        "required": ["status"]
      },
      "TodoListResponse": {
        "type": "object",
        "properties": {
          "todos": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Todo" }
          },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" }
        },
        "required": ["todos", "total", "page", "limit"]
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "List accessible todos",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/TodoStatus" } },
          { "name": "categoryId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "startDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Todo list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TodoListResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "Create todo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Todo" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/todos/today": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get today's todos in KST",
        "responses": {
          "200": { "description": "Todos for today", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Todo" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/todos/this-week": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get this week's todos in KST",
        "responses": {
          "200": { "description": "Todos for this week", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Todo" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/todos/{id}": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get todo by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Todo", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Todo" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "tags": ["Todos"],
        "summary": "Update todo",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated todo", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Todo" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "Delete todo",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/todos/{id}/status": {
      "patch": {
        "tags": ["Todos"],
        "summary": "Update todo status",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoStatusUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Todo" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "422": { "description": "Invalid state transition", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}